+++ core/src/api/BlockchainExplorerEngines.cpp
@@ -7,4 +7,8 @@ namespace ledger { namespace core { namespace api {
 
 std::string const BlockchainExplorerEngines::LEDGER_API = {"LEDGER_API"};
 
+std::string const BlockchainExplorerEngines::RIPPLE_API = {"RIPPLE_API"};
+
+std::string const BlockchainExplorerEngines::RIPPLE_NODE = {"RIPPLE_NODE"};
+
 } } }  // namespace ledger::core::api