+++ core/src/bitcoin/BitcoinLikeExtendedPublicKey.hpp
@@ -31,28 +31,30 @@
 #ifndef LEDGER_CORE_BITCOINLIKEEXTENDEDPUBLICKEY_HPP
 #define LEDGER_CORE_BITCOINLIKEEXTENDEDPUBLICKEY_HPP
 
-#include "../api/BitcoinLikeExtendedPublicKey.hpp"
-#include "../crypto/DeterministicPublicKey.hpp"
-#include "../api/BitcoinLikeNetworkParameters.hpp"
 #include <memory>
-#include "../utils/Option.hpp"
-#include "../utils/DerivationPath.hpp"
+#include <common/AbstractExtendedPublicKey.h>
+#include <api/BitcoinLikeExtendedPublicKey.hpp>
+#include <crypto/DeterministicPublicKey.hpp>
+#include <api/BitcoinLikeNetworkParameters.hpp>
+#include <utils/Option.hpp>
+#include <utils/DerivationPath.hpp>
 #include <api/Currency.hpp>
 
 namespace ledger {
     namespace core {
-        class BitcoinLikeExtendedPublicKey : public api::BitcoinLikeExtendedPublicKey {
+        using BitcoinExtendedPublicKey = AbstractExtendedPublicKey<api::BitcoinLikeNetworkParameters>;
+        class BitcoinLikeExtendedPublicKey : public BitcoinExtendedPublicKey, public api::BitcoinLikeExtendedPublicKey {
         public:
             BitcoinLikeExtendedPublicKey(const api::Currency& params,
                                          const DeterministicPublicKey& key, const DerivationPath& path = DerivationPath("m/"));
-            virtual std::shared_ptr<api::BitcoinLikeAddress> derive(const std::string &path) override;
-            virtual std::shared_ptr<BitcoinLikeExtendedPublicKey> derive(const DerivationPath& path);
+            std::shared_ptr<api::BitcoinLikeAddress> derive(const std::string &path) override;
+            std::shared_ptr<BitcoinLikeExtendedPublicKey> derive(const DerivationPath& path);
 
             std::vector<uint8_t> derivePublicKey(const std::string &path) override;
 
             std::vector<uint8_t> deriveHash160(const std::string &path) override;
 
-            virtual std::string toBase58() override;
+            std::string toBase58() override;
 
             std::string getRootPath() override;
 
@@ -71,8 +73,20 @@ namespace ledger {
                 const Option<std::string>& path
             );
 
+        protected:
+            const api::BitcoinLikeNetworkParameters &params() const override {
+                return _currency.bitcoinLikeNetworkParameters.value();
+            };
+            const DeterministicPublicKey &getKey() const override {
+                return _key;
+            };
+            const DerivationPath &getPath() const override {
+                return _path;
+            };
+            const api::Currency &getCurrency() const override {
+                return _currency;
+            };
         private:
-            inline const api::BitcoinLikeNetworkParameters& params() const;
 
             const api::Currency _currency;
             const DerivationPath _path;